+++ elm.json
@@ -5,6 +5,7 @@
         "./packages/component-library/stories",
         "./packages/hosted-assets",
         "./packages/design-tokens/elm",
+        "./packages/notification/src",
         "./packages/progress-bar/src",
         "./draft-packages/button",
         "./draft-packages/card",
