+++ packages/component-library/components/index.ts
@@ -3,7 +3,6 @@ export * from "./Layout"
 export * from "./NavigationBar"
 export * from "./Icon"
 export * from "./MenuList"
-export * from "./Notification"
 export * from "./OffCanvas"
 export * from "./Text"
 export * from "./Spacing"
