+++ packages/notification/docs/ElmGlobalNotification.stories.tsx
@@ -1,8 +1,8 @@
 import { loadElmStories } from "elm-storybook"
 import { CATEGORIES } from "../../../storybook/constants"
 
-const compiledElm = require("../ElmStories/GlobalNotificationStories.elm").Elm
-  .ElmStories.GlobalNotificationStories
+const compiledElm = require("../src/ElmStories/GlobalNotificationStories.elm")
+  .Elm.ElmStories.GlobalNotificationStories
 
 loadElmStories(`${CATEGORIES.elm}/Global Notification`, module, compiledElm, [
   "Positive",
