+++ packages/notification/docs/ElmToastNotification.stories.tsx
@@ -1,8 +1,8 @@
 import { loadElmStories } from "elm-storybook"
 import { CATEGORIES } from "../../../storybook/constants"
 
-const compiledElm = require("../ElmStories/ToastNotificationStories.elm").Elm
-  .ElmStories.ToastNotificationStories
+const compiledElm = require("../src/ElmStories/ToastNotificationStories.elm")
+  .Elm.ElmStories.ToastNotificationStories
 
 loadElmStories(`${CATEGORIES.elm}/Toast Notification`, module, compiledElm, [
   "Positive",
