+++ packages/notification/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@kaizen/notification",
+  "version": "0.0.0",
+  "description": "Collection of notification components",
+  "scripts": {
+    "prepublish": "tsc --project tsconfig.dist.json",
+    "build": "yarn clean && yarn prepublish",
+    "build:watch": "yarn clean && yarn prepublish --watch",
+    "clean": "rimraf '**/*.d.ts' '**/*.js' '**/*.map'"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cultureamp/kaizen-design-system.git",
+    "directory": "packages/notification"
+  },
+  "bugs": {
+    "url": "https://github.com/cultureamp/kaizen-design-system/issues"
+  },
+  "files": [
+    "**/*",
+    "!**/*.ts",
+    "!**/*.tsx",
+    "**/*.d.ts",
+    "!**/*.spec.*",
+    "!**/*.snap",
+    "!draft-templates",
+    "!stories",
+    "!docs",
+    "!tsconfig.dist.json"
+  ],
+  "author": "",
+  "private": false,
+  "license": "MIT",
+  "dependencies": {
+    "@kaizen/component-library": "^10.3.1",
+    "@kaizen/react-deprecate-warning": "^1.1.5",
+    "classnames": "^2.3.1",
+    "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "@kaizen/design-tokens": "^3.0.0",
+    "rimraf": "^3.0.2"
+  },
+  "peerDependencies": {
+    "@kaizen/design-tokens": "^2.10.3 || ^3.0.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0"
+  }
+}
