+++ packages/notification/src/ElmStories/ToastNotificationStories.elm
@@ -1,63 +1,64 @@
-module ElmStories.GlobalNotificationStories exposing (main)
+module ElmStories.ToastNotificationStories exposing (main)
 
 import ElmStorybook exposing (statelessStoryOf, storybook)
 import Html exposing (Html, a, div, text)
-import Html.Attributes exposing (href, style)
-import Notification.Notification as Notification
+import Html.Attributes exposing (href)
+import Kaizen.Notification as Notification
+import Text.Text as Text
 
 
 main =
     storybook
         [ statelessStoryOf "Positive" <|
             Notification.view
-                (Notification.global [ text "New user data, imported by mackenzie@hooli.com has successfully uploaded. ", a [ href "/" ] [ text "Manage users is now available" ] ]
+                (Notification.toast "Success" [ text "New user data, imported by mackenzie@hooli.com has successfully uploaded. ", a [ href "/" ] [ text "Manage users is now available" ] ] False
                     |> Notification.notificationType Notification.Affirmative
                 )
                 (Notification.Manual Notification.Visible)
                 (always "")
         , statelessStoryOf "Informative" <|
             Notification.view
-                (Notification.global [ text "New user data is currently being processed. We'll let you know when the process is completed. ", a [ href "/" ] [ text "Manage users" ] ]
+                (Notification.toast "Informative" [ text "New user data is currently being processed. We'll let you know when the process is completed. ", a [ href "/" ] [ text "Manage users" ] ] False
                     |> Notification.notificationType Notification.Informative
                 )
                 (Notification.Manual Notification.Visible)
                 (always "")
         , statelessStoryOf "Cautionary" <|
             Notification.view
-                (Notification.global [ text "New user data, imported by mackenzie@hooli.com has uploaded with some minor issues. ", a [ href "/" ] [ text "View issues" ] ]
+                (Notification.toast "Warning" [ text "New user data, imported by mackenzie@hooli.com has uploaded with some minor issues. ", a [ href "/" ] [ text "View issues" ] ] False
                     |> Notification.notificationType Notification.Cautionary
                 )
                 (Notification.Manual Notification.Visible)
                 (always "")
         , statelessStoryOf "Negative" <|
             Notification.view
-                (Notification.global [ text "No network connection. Check your connection and try again. ", a [ href "/" ] [ text "Refresh" ] ]
+                (Notification.toast "No network connection" [ text "Check your connection and try again. ", a [ href "/" ] [ text "Refresh" ] ] False
                     |> Notification.notificationType Notification.Negative
                 )
                 (Notification.Manual Notification.Visible)
                 (always "")
         , statelessStoryOf "Multiple notifications" <|
-            div [ style "width" "inherit" ]
+            div []
                 [ Notification.view
-                    (Notification.global [ text "New user data, imported by mackenzie@hooli.com has successfully uploaded. ", a [ href "/" ] [ text "Manage users is now available" ] ]
+                    (Notification.toast "Success" [ text "New user data, imported by mackenzie@hooli.com has successfully uploaded. ", a [ href "/" ] [ text "Manage users is now available" ] ] False
                         |> Notification.notificationType Notification.Affirmative
                     )
                     (Notification.Manual Notification.Visible)
                     (always "")
                 , Notification.view
-                    (Notification.global [ text "New user data is currently being processed. We'll let you know when the process is completed. ", a [ href "/" ] [ text "Manage users" ] ]
+                    (Notification.toast "Informative" [ text "New user data is currently being processed. We'll let you know when the process is completed. ", a [ href "/" ] [ text "Manage users" ] ] False
                         |> Notification.notificationType Notification.Informative
                     )
                     (Notification.Manual Notification.Visible)
                     (always "")
                 , Notification.view
-                    (Notification.global [ text "New user data, imported by mackenzie@hooli.com has uploaded with some minor issues. ", a [ href "/" ] [ text "View issues" ] ]
+                    (Notification.toast "Warning" [ text "New user data, imported by mackenzie@hooli.com has uploaded with some minor issues. ", a [ href "/" ] [ text "View issues" ] ] False
                         |> Notification.notificationType Notification.Cautionary
                     )
                     (Notification.Manual Notification.Visible)
                     (always "")
                 , Notification.view
-                    (Notification.global [ text "No network connection. Check your connection and try again. ", a [ href "/" ] [ text "Refresh" ] ]
+                    (Notification.toast "No network connection" [ text "Check your connection and try again. ", a [ href "/" ] [ text "Refresh" ] ] False
                         |> Notification.notificationType Notification.Negative
                     )
                     (Notification.Manual Notification.Visible)
