+++ packages/notification/src/ToastNotificationsList.module.scss
