+++ packages/notification/src/components/GenericNotification.module.scss
