+++ packages/notification/src/components/GenericNotification.tsx
@@ -1,11 +1,11 @@
 import classnames from "classnames"
 import * as React from "react"
-import { Icon } from "../../Icon"
+import { Icon } from "@kaizen/component-library"
 
-import closeIcon from "../../../icons/close.icon.svg"
-import exclamationIcon from "../../../icons/exclamation.icon.svg"
-import informationIcon from "../../../icons/information.icon.svg"
-import successIcon from "../../../icons/success.icon.svg"
+import closeIcon from "@kaizen/component-library/icons/close.icon.svg"
+import exclamationIcon from "@kaizen/component-library/icons/exclamation.icon.svg"
+import informationIcon from "@kaizen/component-library/icons/information.icon.svg"
+import successIcon from "@kaizen/component-library/icons/success.icon.svg"
 
 import styles from "./GenericNotification.module.scss"
